+++ srcs/backend/transcendence/settings.py
@@ -44,7 +44,6 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-	'ted',
 	'user',
 	'game',
 	'tournament',
