+++ srcs/backend/user/views.py
@@ -1,3 +1,12 @@
 from django.shortcuts import render
+from django.http import HttpResponse
+from django.template import loader
 
-# Create your views here.
+def index(request):
+	context = {
+		"message": "PONG",
+		"number": 42,
+		"userList": ["Yabing", "Ludo", "David", "Thibaud"]
+		}
+	template = loader.get_template("../frontend/index.html")
+	return HttpResponse(template.render(context, request))
