+++ srcs/frontend/css/global.css
